refactor: centralize Firebase initialization
