+++ src/pages/LoginHistory.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useAuth } from "../contexts/AuthContext";
 import { useToast } from "../contexts/ToastContext";
-import { db } from "../utils/firebase";
+import { db } from "../config/firebase";
 import {
   collection,
   query,
